573: Move ParseLoggingQuery function and constant from controller to client
