+++ api/controllers/service.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 
 	"github.com/quintilesims/layer0/api/provider"
+	"github.com/quintilesims/layer0/client"
 	"github.com/quintilesims/layer0/common/errors"
 	"github.com/quintilesims/layer0/common/models"
 	"github.com/zpatrick/fireball"
@@ -93,7 +94,7 @@ func (s *ServiceController) readService(c *fireball.Context) (fireball.Response,
 
 func (s *ServiceController) readServiceLogs(c *fireball.Context) (fireball.Response, error) {
 	serviceID := c.PathVariables["id"]
-	tail, start, end, err := ParseLoggingQuery(c.Request.URL.Query())
+	tail, start, end, err := client.ParseLoggingQuery(c.Request.URL.Query())
 	if err != nil {
 		return nil, errors.New(errors.InvalidRequest, err)
 	}
