+++ api/controllers/task_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/golang/mock/gomock"
 	"github.com/quintilesims/layer0/api/provider/mock_provider"
+	"github.com/quintilesims/layer0/client"
 	"github.com/quintilesims/layer0/common/models"
 	"github.com/stretchr/testify/assert"
 )
@@ -159,12 +160,12 @@ func TestReadTaskLogs(t *testing.T) {
 	}
 
 	tail := "100"
-	start, err := time.Parse(TimeLayout, "2001-01-02 10:00")
+	start, err := time.Parse(client.TimeLayout, "2001-01-02 10:00")
 	if err != nil {
 		t.Fatalf("Failed to parse start: %v", err)
 	}
 
-	end, err := time.Parse(TimeLayout, "2001-01-02 12:00")
+	end, err := time.Parse(client.TimeLayout, "2001-01-02 12:00")
 	if err != nil {
 		t.Fatalf("Failed to parse end: %v", err)
 	}
@@ -176,8 +177,8 @@ func TestReadTaskLogs(t *testing.T) {
 	c := newFireballContext(t, nil, map[string]string{"id": "tsk_id"})
 	c.Request.URL.RawQuery = fmt.Sprintf("tail=%s&start=%s&end=%s",
 		tail,
-		start.Format(TimeLayout),
-		end.Format(TimeLayout))
+		start.Format(client.TimeLayout),
+		end.Format(client.TimeLayout))
 
 	resp, err := controller.readTaskLogs(c)
 	if err != nil {
