+++ client/admin.go
@@ -3,7 +3,6 @@ package client
 import (
 	"net/url"
 
-	"github.com/quintilesims/layer0/api/controllers"
 	"github.com/quintilesims/layer0/common/models"
 	"github.com/zpatrick/rclient"
 )
@@ -18,7 +17,7 @@ func (c *APIClient) ReadConfig() (*models.APIConfig, error) {
 }
 
 func (c *APIClient) ReadAdminLogs(query url.Values) ([]models.LogFile, error) {
-	if _, _, _, err := controllers.ParseLoggingQuery(query); err != nil {
+	if _, _, _, err := ParseLoggingQuery(query); err != nil {
 		return nil, err
 	}
