+++ client/environment.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net/url"
 
-	"github.com/quintilesims/layer0/api/controllers"
 	"github.com/quintilesims/layer0/common/models"
 	"github.com/zpatrick/rclient"
 )
@@ -51,7 +50,7 @@ func (c *APIClient) ReadEnvironment(environmentID string) (*models.Environment, 
 }
 
 func (c *APIClient) ReadEnvironmentLogs(environmentID string, query url.Values) ([]models.LogFile, error) {
-	if _, _, _, err := controllers.ParseLoggingQuery(query); err != nil {
+	if _, _, _, err := ParseLoggingQuery(query); err != nil {
 		return nil, err
 	}
